+++ fwdpy11/headers/fwdpy11/discrete_demography/simulation/functions.hpp
@@ -60,10 +60,9 @@ namespace fwdpy11
         namespace detail
         {
             inline void
-            update_current_deme_sizes(
-                const std::uint32_t t,
-                current_event_state<SetDemeSize>& size_change_events,
-                deme_properties& sizes_rates)
+            set_next_deme_sizes(const std::uint32_t t,
+                                current_event_state<SetDemeSize>& size_change_events,
+                                deme_properties& sizes_rates)
             // NOTE: this function may resent growth rates to zero.
             // see SetDemeSize for details.
             {
@@ -72,8 +71,8 @@ namespace fwdpy11
                     {
                         return;
                     }
-                current_deme_sizes_vector::value_type& current_deme_sizes
-                    = sizes_rates.current_deme_sizes.get();
+                next_deme_sizes_vector::value_type& current_deme_sizes
+                    = sizes_rates.next_deme_sizes.get();
                 growth_rates_vector::value_type& growth_rates
                     = sizes_rates.growth_rates.get();
                 growth_rates_onset_times_vector::value_type& growth_rate_onset_times
@@ -111,16 +110,15 @@ namespace fwdpy11
                     }
                 auto& rates = sizes_rates.growth_rates.get();
                 auto& onsets = sizes_rates.growth_rate_onset_times.get();
-                auto& current_deme_sizes = sizes_rates.current_deme_sizes.get();
+                auto& next_deme_sizes = sizes_rates.next_deme_sizes.get();
                 auto& N0 = sizes_rates.growth_initial_sizes.get();
-                auto& Ncurr = sizes_rates.current_deme_sizes.get();
                 for (; growth_rate_changes.current() < growth_rate_changes.last()
                        && growth_rate_changes.when() == t;
                      ++growth_rate_changes.current())
                     {
                         auto deme = growth_rate_changes.event().deme;
                         if (growth_rate_changes.event().G != NOGROWTH
-                            && Ncurr[deme] == 0)
+                            && next_deme_sizes[deme] == 0)
                             {
                                 throw DemographyError(
                                     "attempt to change growth rate in extinct "
@@ -128,7 +126,7 @@ namespace fwdpy11
                             }
                         rates[deme] = growth_rate_changes.event().G;
                         onsets[deme] = t;
-                        N0[deme] = current_deme_sizes[deme];
+                        N0[deme] = next_deme_sizes[deme];
                     }
             }