+++ fwdpy11/src/evolve_population/evolvets.cc
@@ -316,6 +316,7 @@ evolve_with_tree_sequences(
             build_migration_lookup(
                 current_demographic_state.M,
                 current_demographic_state.current_deme_parameters.current_deme_sizes,
+                current_demographic_state.current_deme_parameters.next_deme_sizes,
                 miglookup);
             fitness_lookup.update(current_demographic_state.fitness_bookmark);
             ddemog::validate_parental_state(